testing

+++ src/base/TensorRadix16.cu
@@ -65,13 +65,13 @@ __global__ void TensorRadix16(__half* input_data_RE, __half* input_data_IM,
   for(int k=0; k<8; k++){
     int j = k + 8 * inter_warp_id_is_upper_16;
     int buffer_array_id = inter_warp_id_16 + 16 * j;
-    
+
     __half phase =
         __hdiv(__hmul(static_cast<__half>(j * inter_warp_id_16),
                       static_cast<__half>(M_PI)),
                static_cast<__half>(8.0));
-    buffer_RE[buffer_array_id] = cos(phase);
-    buffer_IM[buffer_array_id] = -sin(phase);
+    buffer_RE[buffer_array_id] = cos(static_cast<__half>(phase));
+    buffer_IM[buffer_array_id] = -sin(static_cast<__half>(phase));
   }
 
   //Literal version of dft matrix.
@@ -118,8 +118,8 @@ __global__ void TensorRadix16(__half* input_data_RE, __half* input_data_IM,
                           static_cast<__half>(tmp));
 
     //TO-SELF: test __cosf vs cos accuracy and speed
-    __half twiddle_RE = cos(phase);
-    __half twiddle_IM = -sin(phase);
+    __half twiddle_RE = cos(static_cast<__half>(phase));
+    __half twiddle_IM = -sin(static_cast<__half>(phase));
 
     //Fetch current data once from global memory to use it twice
     //For unscaled or scaling at once
